% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSF.R
\name{CSF}
\alias{CSF}
\title{Airborne LiDAR filtering method based on Cloth Simulation}
\usage{
CSF(cloud, sloop_smooth = FALSE, class_threshold = 0.5,
  cloth_resolution = 0.5, rigidness = 1L, interations = 500L,
  time_step = 0.65)
}
\arguments{
\item{cloud}{data.frame with 3 columns named X Y, Z containing the coordinates of the point cloud.}

\item{sloop_smooth}{logical.}

\item{class_threshold}{scalar.}

\item{cloth_resolution}{scalar.}

\item{rigidness}{integer.}

\item{time_step}{scalar.}

\item{interation}{integer.}
}
\value{
A vector containing the ids of the points that belong on the ground.
}
\description{
Airborne LiDAR filtering method based on Cloth Simulation (Zhang et al. 2016, see references). This
function is an R wrapper around the library written by the orignal authors of the algorithm.
}
\examples{
data(rcsf_cloud)
head(rcsf_cloud)

id_ground = CSF(rcsf_cloud)
}
\references{
W. Zhang, J. Qi*, P. Wan, H. Wang, D. Xie, X. Wang, and G. Yan, “An Easy-to-Use Airborne
LiDAR Data Filtering Method Based on Cloth Simulation,” Remote Sens., vol. 8, no. 6, p. 501, 2016
}
