% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{rcsf_cloud}
\alias{rcsf_cloud}
\title{Airborne LiDAR point cloud}
\format{A data frame with 28668 rows and 3 variables:
\describe{
  \item{X}{x coordinates}
  \item{Y}{y coordinates}
  \item{Z}{z coordinates}
}}
\usage{
rcsf_cloud
}
\description{
A dataset containing a small point cloud aquiered with airborne LiDAR.
}
\keyword{datasets}
