% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSF.R
\name{CSF}
\alias{CSF}
\title{Airborne LiDAR filtering method based on Cloth Simulation}
\usage{
CSF(cloud, SloopSmooth = FALSE, ClothResolution = 0.5)
}
\arguments{
\item{cloud}{data.frame with 3 columns names X Y, Z containing the coordinates of the point cloud.}

\item{SloopSmooth}{logical}

\item{ClothResolution}{numeric.}
}
\value{
A vector containing the ids of the points that belong on the ground.
}
\description{
Airborne LiDAR filtering method based on Cloth Simulation
}
\references{
W. Zhang, J. Qi*, P. Wan, H. Wang, D. Xie, X. Wang, and G. Yan, “An Easy-to-Use Airborne
LiDAR Data Filtering Method Based on Cloth Simulation,” Remote Sens., vol. 8, no. 6, p. 501, 2016
}
